[FIX] product_variants_no_automatic_creation: Find correctly a product for its attributes
Adding to the domain all values that exists, not only the ones that have more than one
possible value. That was an old constraint because attributes with only one value were
not populated, but that was fixed a long time ago and now it makes that the match doesn't
happen because the comparison of the attributes length fail.